Remove the deploy of portfolio_projects pages
We are creating so many pages in the model app
i.e. for each of the portfolio_project, outside_project,
affiliated_committer and outside_committer which causes the
netlify to fail in the next commit.

This commit removes the deploy of portfolio_projects pages
which minimises the number of pages we deploy and the netlify
passes. The details of all the portfolio_projects can also be
view at '/openhub/'.