Remove the deploy of portfolio_projects pages
We do not know why netlify is failing but from testing
we have found out that minimising the number of pages
being deploy by netlify lead to a passing netlify.

So, this commit removes the deploy of portfolio_projects
pages to achieve the same for the time being. The details
of all the portfolio_projects can also be view at '/openhub/'.