git.py : Add name argument to get_upstream_deploy_url method
fetch_deployed_data.py : Pass an argument to the get_upstream_deploy_url

When fetching the deployed data from the upstream, it tried to fetch
from the current repo even if the data is from a different repo. Thus
adding a repo name argument ensures that the data is fetched from the
repo it is actually in.

Fixes https://github.com/coala/community/issues/185