git.py: Make netlify build pass for local branch
When the forked community repo is deployed on netlify
the build is used to fail due to non-availability of
attributes name, owner and resource in url attribute.
They're used to have None value instead of a valid
value. So, to resolve this issue I've added a way to
get valid values of these attrs from href attribute in
url object.

Fixes https://github.com/coala/community/issues/232