ci_build/: Re-design build-logs webpage
The newly created webpage combines the previous
two webpages- info.txt and log/index.html. This
web-page combines the results of both the pages
and shows them in a better UI/UX with additional
features of filtering and searching within the
existing logs. The logs are fetched from a JSON
file which is created from the logs stored in
the log file _site/community.log

Closes https://github.com/coala/community/issues/256