community/: Add a joining netlify form
The netlify form will ask some particular
inputs that will be used for validating
the user - whether the user is eligible
to be a organization member. The checks for
it have been defined in coala webservices API
which will be accepting form-submissions over
a cron-job defined. After the user submits the
form, a success message will be displayed.

Closes https://github.com/coala/community/issues/89, https://github.com/coala/community/issues/266