community/: Display contributor statistics
Show the contributor statistics in the form of
charts. The charts will be shown for issues,
commits, reviews and for merge requests. And,
an overall chart, which will display stats
repository wise. Apart from all these charts,
the type of issues the contrib has worked on
will be shown, with the count of issues the
user is working on and the last updated datetime
the user data was updated. All this will be shown,
only if there is some statistics, otherwise the
popup option willn't be shown.

Closes https://github.com/coala/community/issues/268