community/: Add a form for adding calendar events
Not everyone, will be able to fill forms.
Only the logged in users will be able to
fill them and some of the forms, can only
be filled by developers or contributors
who are a part of more than one team.
At every step, the check is performed whether
the user is authenticated or not, to avoid
false form submissions.

Closes https://github.com/coala/community/issues/270