community/: Add a webpage for Listing Issues
This commit adds a general view for listing
all the issue objects for any model which
stores issue related details. The details
which are compulsory needed are Hoster,
Repository name, Issue title, Number and
the URl of it. using this, the commit adds
a webpage for displaying all the inactive
issues.

Closes https://github.com/coala/community/issues/288